THE FLU UPDATE IS HERE
Sooo I decided to write code to give our robot the flu. So, I present to you, the long awaited flu update. Featuring both the A and B strand of the flu! There is a method that will cure the A strand, but not the B strand. (I have no idea why I wasted time on this). Enjoy?!?!